*** CODE REVIEW REQUEST *** Fixed issues with long path names: 1) Long path names were not correctly watched 2) Sometimes crashed due to buffer overrun. Two alternatives are for review: a) solution with static buffer and b) solution with dynamically-allocated buffer (depending on value of USE_STATIC_BUF.
I will do the clone cleanup based on your feedback and keep only the solution you'll select.

Also, I do understand why there is a reference to std::wstring on line 251. It allocates a new std::wstring object just for the call to the .insert method. Could buf_ptr be used directly as argument to the .insert method?