Last accessed timestamp is modified when we reset the source file's archive attribute.
To prevent this change, cryptsync was restoring the source file's timestamp after the attribut change. This trigger Window to notify of a source file change.

The benefit of preseving the source file's access time is limited so the code was removed.

Furthermore, the last access time of the file is recorded in the .7z file. Since the 7z library does not call SetFileTime after the source file is open with 0xFFFFFFFF, the timestamp recorded is the time cryptsync read the file.

Also, when decrypting the file either with 7Zip or cryptsync, the creation and last access time are those of the extraction.

The extra Wndows change notification has no functional impact.